% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ANOVAShiny2.R
\name{ANOVAShiny2}
\alias{ANOVAShiny2}
\title{Launch 'ANOVAShiny2' Interface}
\usage{
ANOVAShiny2()
}
\value{
Nothing
}
\description{
ANOVAShiny2() loads interactive user interface built using R 'shiny'.
}
\details{
Start ANOVAShiny2

The interactive user interface is to provide an easy way for anova analysis and downloading relevant plots.
}
\examples{
if(interactive()){
library(rmarkdown)
ANOVAShiny2()
}
}
\keyword{ANOVAShiny2}
